Split up hooks into separate modules
To distinguish the hookspec categories:

- csp_hookspecs for csp and billing interaction
- storage_hookspecs for CRUD of cache and CSP config
- hookspecs for everything else
